+++ bidsme/Modules/MRI/NIFTI.py
@@ -131,7 +131,7 @@ class NIFTI(MRI):
             res = None
         return res
 
-    def copyRawFile(self, destination: str) -> None:
+    def copyRawFile(self, destination: str) -> str:
         if os.path.isfile(os.path.join(destination,
                                        self.currentFile(True))):
             logger.warning("{}: File {} exists at destination"
@@ -141,6 +141,7 @@ class NIFTI(MRI):
         if self._nii_type == "ni1":
             data_file = tools.change_ext(self.currentFile(), "img")
             shutil.copy2(data_file, destination)
+        return os.path.join(destination, self.currentFile(True))
 
     def _copy_bidsified(self, directory: str, bidsname: str, ext: str) -> None:
         if self._nii_type == "ni1":