+++ bidsme/Modules/PET/ECAT.py
@@ -26,7 +26,6 @@ from .PET import PET
 from ..common import action_value
 from . import _ECAT
 
-import os
 import logging
 import numpy
 
@@ -42,6 +41,7 @@ class ECAT(PET):
     __slots__ = ["_ECAT_CACHE", "_SUB_CACHE", "_FILE_CACHE"]
     __specialFields = {"ScanStart", "InjectionStart",
                        "FramesStart", "FramesDuration"}
+    _file_extentions = [".v"]
 
     def __init__(self, rec_path=""):
         super().__init__()
@@ -71,22 +71,15 @@ class ECAT(PET):
         bool:
             True if file is identified as ECAT
         """
-        if not os.path.isfile(file):
-            return False
-        if file.endswith(".v"):
-            if os.path.basename(file).startswith('.'):
-                logger.warning('{}: file {} is hidden'
-                               .format(cls.formatIdentity(),
-                                       file))
-            try:
-                with open(file, "rb") as f:
-                    magic = f.read(14).decode()
-                    magic = magic.strip(" \0")
-                    if magic.startswith("MATRIX"):
-                        return True
-            except Exception:
+        with open(file, "rb") as f:
+            magic = f.read(14).decode()
+            magic = magic.strip(" \0")
+            if magic.startswith("MATRIX"):
+                return True
+            else:
+                logger.debug("{}: Missing magic string"
+                             .format(cls.formatIdentity()))
                 return False
-        return False
 
     def _loadFile(self, path: str) -> None:
         if path != self._FILE_CACHE: